Move package into a sub-directory of /packages
