+++ packages/package-name/jest.config.js
