+++ packages/package-name/package.json
