+++ packages/package-name/src/index.spec.ts
