+++ packages/package-name/src/index.ts
